XD-1144 Added basic support for module colocation in DeploymentManifest
The DeploymentManifest can specify for a specific stream the names of modules
that it should colocate via a colocationGroupName.  The transformation will
set the group name on matching ModuleDeploymentRequests.  It will be up to
the deployer to use this group name to send the module to the same node.

Note that this does not create a composite module out of modules that appear
sequentially in the stream processing pipeline.